# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AbacusAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/subscription/create:
    post:
      operationId: createSubscription
      tags: [subscription]
      summary: Create a subscription order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSubscriptionResponse"
  /v1/subscription/webhook:
    post:
      operationId: subscriptionWebhook
      tags: [subscription]
      summary: Razorpay webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/subscription/verify:
    post:
      operationId: verifyPayment
      tags: [subscription]
      summary: Verify Razorpay payment and activate subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPaymentResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    CreateSubscriptionBody:
      type: object
      properties:
        plan:
          type: string
          enum: [STAR, SCHOOL]
        billingCycle:
          type: string
          enum: [monthly, yearly]
      required:
        - plan
        - billingCycle
    CreateSubscriptionResponse:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
      required:
        - orderId
        - amount
        - currency
        - keyId
    VerifyPaymentBody:
      type: object
      properties:
        razorpay_payment_id:
          type: string
        razorpay_order_id:
          type: string
        razorpay_signature:
          type: string
        plan:
          type: string
          enum: [STAR]
        billingCycle:
          type: string
          enum: [monthly, yearly]
      required:
        - razorpay_payment_id
        - razorpay_order_id
        - razorpay_signature
        - plan
        - billingCycle
    VerifyPaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        plan:
          type: string
      required:
        - success
        - plan
...